safari-534.53-branch/LayoutTests/platform/mac/accessibility/visible-position-crash-for-text-node-expected.txt

This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text. This is a lot of text.
This protects against a crash in asking for visible point for position to ensure that it can handle all RenderText objects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS content.childAtIndex(0).isEqual(textElement) is true
PASS successfullyParsed is true

TEST COMPLETE
